Improve file import controller specs
Unfortunately, the JS part of the importer feature was written using
Backbone's `fetch` with a custom `type` parameter instead of using
`sync`. This causes the request to use a form data content type
instead of JSON. Consequently, the arrays of selected files are
represented as hashes of the form `{'0' => ...}` instead of arrays.

Even more unfortunately, this sub-optimal representation then also
became part of the contract of `FileImporter#files_meta_data` passing
the `ActionController::Params` object wrapping this hash along instead
of an array.

To prevent breaking integrations in future changes, we now capture
this reality more closely in the specs. We unify specs to all use this
representation and can thus remove defensive handling of file arrays
in the file importer.

[1] https://github.com/codevise/pageflow/blob/54461defce06893bb02e20514fb15d6fc0121a1c/package/src/editor/models/FileImport.js#L71