Use file partial to render start_import_job response
Responding with an incomplete JSON representation of files leads to
problems. Especially the missing perma_id attribute causes the file to
break collection syncing of Pageflow Next. Since the id attribute is
present, the file is included in collections. Since collections are
indexed by perma_id, an invalid item with undefined key is
created. This is then looked up as the referenced file, when actually
no file has been selected.

REDMINE-19898