Remove space before heading in cards appearance
We need to find a good alternative to the fixed space before headings
in the first section. Since some entries rely on the space, though, we
cannot simply remove it.

At least when the cards are used, we know that we do not want to add a
padding to the card. So we start by fixing this case as a special
case.

REDMINE-19912