Allow adding consent opt-in story for content elements
Allow rendering content elements the way they look before consent has
been given in storybook. So far the storybook relies on the fact, that
consent is only required when a consent cookie name has been
configured. Since this is not configured by default, consent opt-in is
never displayed in the storybook.

Content elements which interact with consent, can now pass the
`consent` option when adding content element stories. We then set a
consent cookie name when constructing the consent object used in an
additional "Consent - Opt In" story.