Switch between static/dynamic shadow via CSS
Before, GradientShadow used the same div to display the static
gradient shadow and the dynamic, faded-in shadow. We now split these
into two divs that can be hidden individually via CSS.

For now, we make the component work with both v1 and v2 frontend code
- also controlling shadow visibility in JS via `isContentPadded`. v2
does not need to provide this information to JS, though.

REDMINE-19888