Ensure inline editing help texts/prompts use user locale
In #1867 we ensured that the `lang` attribute on the `html` tag is set
based on the entry's locale in the editor. By switching the locale for
the whole seed template, we also made inline editing help texts and
prompts use the entry locale. Only set `lang` attribute based on entry
locale instead.

REDMINE-19967