Allow looking up published entries via permalinks
We want to give users more freedom when it comes to designing the URLs
of their entries. Currently, entry slugs have to be unique across all
accounts. Since we depend on slugs like that for admin URLs and some
other use cases, we introduce permalinks as an additional top-level
concept.

We also introduce permalink directories as a way to build more complex
sitemaps. Permalink slugs only need to be unique within there
directories.

* Requiring each permalink to have a directory lets us roll out the
  feature step by step, by manually creating permalink directories for
  specific accounts.

* Giving themings root permalink directories ensures that there can be
  no longer any naming conflicts between accounts.

* Two entries within the same theming can use the same slug as long as
  they are in different directories (e.g. /de/start and /en/start).

REDMINE-19973