Set default permalink slug when empty string
We display a default slug based on the entered entry title as a
placeholder in the permalink slug input. In case the default slug is
already taken within the selected permalink directory, we have no good
way to predict the slug Friendly ID would generate in this case (UUIDs
will be different each time). Also, now that the permalink is made
explicit as a concept, we assume that it will be less confusing to see
an error message that the slug is already taken than have a random
looking string appended to the slug.

When the form is submitted with an empty slug, we thus assign the
default slug and render a validation error if it already taken.

In contrast, when duplicating an entry, the slug will be nil causing
Friendly ID to auto-generate a slug based on the entry title.

REDMINE-19973