Include more info in events used for tracking scrolled entries
So far, Scrolled mimicked the payload of Paged's `page:change` and
media event payload. We now want to provide richer data as part of
these events. Since Paged uses the actual page jQuery widget instance
as payload of the `page:change` event, the easiest way to provide
additional data there is via a new method on the instance. Going
forward, this method will be the shared interface of events triggered
by Paged and Scrolled.

REDMINE-19477