Disable top padding for new headings
The automatic padding for headings in first sections keeps confusing
users. Still, we cannot simply remove it without risking breaking
existing entries. We therefore add a new default config which turns it
off for new headings. Since we expect to add some sort of margin
settings for all content element types at some point, we try to name
the property in a way that we can continue to use once we introduce
this concept.

REDMINE-20084