Prevent foreground shadow from overlaying next section
For sections that exit with a fade transition we extend the height of
the foreground shadow to make sure the shadow does not end too early
before the section is finally faded out.

When the "use full viewport" setting is disabled for such a section,
we fall back to the "scroll" transition. In this case we still made
the shadow longer causing it cover parts of the following section.