Automatically abort file selection
When creating file links it is currently hard to for the link
destination dialog to figure out when the user decides to navigate
away from the file list instead of finishing the process. The current
link destination selection workflow depends on aborting the process,
though, to prevent inserting another link when the process is started
again and the previous selection promise is still pending.

Instead, we can abort the process automatically, whenever it is
started anew.

REDMINE-20140