Do not scale images up when creating panorama styles
Especially for small images with extreme aspect ratios, this can lead
to images which exceed limits and cause errors of the form:

    convert-im6.q16: width or height exceeds limit

For example, a 1000x10 image scaled to cover 1024x1024 a rect leads to
an image that is 102400 pixels wide.

Instead only scale panorama images if their dimensions exceed the size
of the box that we want to cover in both dimensions.

To be able to use image dimensions to define style geometry, we need
to detect width and height before post processing starts.

REDMINE-20147