Display rights of all files in credits box
Previously only rights of image files where displayed. To do this, we
need to extract rights information from all files collections. Add a
new `useMultipleEntryStateCollectionItems` hook based on a new
`createMultipleItemsSelector` function, which can be used to extract
multiple collections at once and only re-render when an item of one of
the collection changes.

REDMINE-20295