Support scopes in storybook theme properties
So far theme properties in stories were set directly as keys below
`themeOptions.properties`. In real themes, a scope like `root` or
`defaultNavigation` needs to be inserted as an intermediate level.

To allow testing these scopes in stories, we now generate a full style
tag with scope rules instead of only setting the properties as inline
styles on a wrapper div.

REDMINE-20411