Prevent normalizing color input too early
When pausing after typing the first three digits of a color
code (e.g., #abc), minicolors already updates the value with a
normalized representation (e.g., #aabbcc). Before, the `change`
handler updated the model, which in turn triggered the `load` method,
causing the value of the input to be updated. This prevented the user
from finishing typing the full color code.

REDMINE-20413