Set inverse of draft and published_revision associations on entry
This enables an edge-case where the account of an example entry is
changed transiently before creating a theme customization
preview. Make sure feature flags of the account are taken into account
correctly.

Generally, this should be an improvement since draft and published
revisions no longer need to look up there entries in a separate query.

REDMINE-20474