Add active flag to media query hook
We intend to pull up the logic that decides which files to render in
the backdrop into a hook on section to make file available for inline
rights display. So far, we prevented creating unneeded orientation
media queries by rendering `OrientationAwareBackgroundImage/Video`
components that used the media query hook only when portrait files
were present.

We now need a way to be able to always call the
`usePortraitOrientation` hook, but prevent it from creating the media
query if no portrait files have been configured.

REDMINE-20093