Store previous src for media event file names
When a player is unallocated, we reset src to a blank media file via a
placeholder data url and pause the player. The player still triggers
timeupdate and pause events that we want to turn into media events. We
do not want to report the data url as file name, since analytics
integrations rely on stable file names. We thus store the previous src
and use that wehn reporting the remaining events.

REDMINE-20591