Support rendering content element as backdrop
Turn `useSectionContentElements` hook into
`useSectionForegroundcontentelements` hook which filters out elements
with position `backdrop`. Introduce `useContentElement` hook to look
up elements by perma id and use it to render content element in the
backdrop if `section.backdrop.contentElement` is set.

At the moment, perma ids might get re-used once an element is deleted
and another one is created. We need to prevent the new element from
being treated as the backdrop of the section if it belongs to another
section. We therefore check the sectionId of the element.

REDMINE-20634