Replace jQuery UI sortable with SortableJS
Less buggy when it comes to sorting items within a scrolling
container.

SortableJS triggers callbacks in different order: `onRemove` is
invoked after `onAdd`. Since removing a model from a
`ForeignKeySubsetCollection` resets the parent reference, we have to
make sure to first remove and than add. Use the fact that - after
`onAdd` - `onSort` is called on both sortables when moving an item and
use it instead of `onRemove`.

SortableJS leaves the dragged element in place and only applies a CSS
class. Use this to simplify the placeholder setup.

REDMIME-20837