Prevent reflow when selecting section
Selecting the section after a pan-zoom hotspot had been selected,
caused the pan-zoom animation to stay stuck in Firefox instead of
reverting to the initial step.

The `mousedown` on the section triggers both the selection change and
the `onDismiss` callback of the floating tooltip that is supposed to
scroll the hotspots scroller back to the first position.

When the section is selected, we toggle a `selected` CSS class on the
section to display a selection rect made of pseudo elements. Toggling
these pseudo elements causes a reflow that makes Firefox drop the
`scroll` event that would have been triggered by scrolling the
hotspots scroller.

The scroll timeline polyfill then has no opportunity to refresh the
position of the hotspots image causing it to stay stuck.

REDMINE-20673