Fit two S/XS items next to each other in full-width scroller
When scroller is active, we apply a padding to the list to replicate
content margin as custom margin. Reference width was calculated via
custom property on list referring to list width 100%. When using that
to compute item width, the 100% is interpreted relative to
the (smaller) content box of the list. Trying to calculate the width
of the content margin thus leads to a too small value, causing the
items to be too large.

REDMINE-20981