Fix outside press dismissal of hotspot tooltips
floating-ui appears to use native event handlers for detecting
mousedown on the document, but React event callbacks for intercepting
mousedown inside the floating element in the capture phase. Looks like
capture phase for React events happens after native events in React
16. React 17 changes the event system here.

This causes the first click on a link tooltip of the hotspot tooltip
link button to close the tooltip before the link can be clicked in the
editor. In the published editor it disables outside press dismissal
once after a click inside the floating element.

Disable capture for outside press to work around these issue.